+++ client/src/components/Card/Card.jsx
@@ -1,14 +1,11 @@
 import React, { useContext } from "react";
 import { useNavigate } from "react-router-dom";
-import addProduct from "../../firebase/addProduct";
 import { userAuth } from "../../context/Auth-context";
 import Button from "../UI/Button";
-import increase from "../../firebase/increase";
-import axios from "axios";
+
 const Card = ({ data }) => {
   const navigate = useNavigate();
-  const { currentUser, setProductsLocalStorage, productsLocalStorage } =
-    useContext(userAuth);
+
   const {
     brand: { name: brandName },
     tipo: { name: tipoName },
@@ -17,69 +14,6 @@ const Card = ({ data }) => {
     id,
     precio,
   } = data;
-
-  const addProducto = async () => {
-    const response = await axios.get(`motos/${id}`);
-    const stock = response.data.stock;
-
-    if (currentUser) {
-      // Si el usuario está autenticado, verifica si el producto ya está en el carrito
-      const existingProduct = productsLocalStorage.find(
-        (product) => product.id === id
-      );
-
-      if (existingProduct) {
-        // Si el producto ya está en el carrito, aumenta la cantidad
-        existingProduct.cantidad += 1;
-        setProductsLocalStorage([...productsLocalStorage]);
-      } else {
-        // Si el producto no está en el carrito, agrégalo
-        addProduct(currentUser.uid, {
-          brand: brandName,
-          tipo: tipoName,
-          imageUrl: imageUrl[0],
-          motoModel,
-          id,
-          precio,
-          cantidad: 1,
-        });
-      }
-
-      increase(currentUser.uid, id);
-    } else {
-      // Si el usuario no está autenticado, verifica si el producto ya está en el carrito local
-      const existingProducts =
-        JSON.parse(localStorage.getItem("products")) || [];
-
-      const existingProduct = existingProducts.find(
-        (product) => product.id === id
-      );
-
-      if (existingProduct) {
-        // Si el producto ya está en el carrito local, aumenta la cantidad
-        existingProduct.cantidad += 1;
-
-        // Actualiza el localStorage y el estado local
-        localStorage.setItem("products", JSON.stringify(existingProducts));
-        setProductsLocalStorage(existingProducts);
-      } else {
-        // Si el producto no está en el carrito local, agrégalo
-        existingProducts.push({
-          brand: brandName,
-          tipo: tipoName,
-          imageUrl: imageUrl[0],
-          motoModel,
-          id,
-          precio,
-          cantidad: 1,
-        });
-
-        // Actualiza el localStorage y el estado local
-        localStorage.setItem("products", JSON.stringify(existingProducts));
-        setProductsLocalStorage(existingProducts);
-      }
-    }
-  };
 
   return (
     <article className=" bg-[#0006]   p-2  rounded-2xl">
@@ -110,11 +44,11 @@ const Card = ({ data }) => {
             </p>
           </section>
           <section className="flex justify-center max-md:flex-col gap-3">
-            <Button
+            {/* <Button
               onClick={addProducto}
               text="Add"
               className=" text-white hover:text-black hover:bg-white text-xl shadow-none "
-            />
+            /> */}
 
             <Button
               onClick={() => navigate(`/detail/${id}`)}
